requirements.txt
fastapi
uvicorn
python-dotenv
swisseph
supabase
httpx
pydantic
python-multipart
pytz
